Remove inspection table
