+++ backend/api/Controllers/MissionSchedulingController.cs
@@ -74,8 +74,7 @@ namespace Api.Controllers
             foreach (var task in missionTasks)
             {
                 task.Id = Guid.NewGuid().ToString();
-                if (task.Inspection != null)
-                    task.Inspection.Id = Guid.NewGuid().ToString();
+                task.IsarInspectionId = Guid.NewGuid().ToString();
             }
 
             var newMissionRun = new MissionRun
