+++ backend/api/Database/Models/MissionTask.cs
@@ -1,13 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
 using Api.Services.Models;
+using Microsoft.EntityFrameworkCore;
 #pragma warning disable CS8618
 namespace Api.Database.Models
 {
     public class MissionTask
     {
-        private TaskStatus _status;
-
         // ReSharper disable once NotNullOrRequiredMemberIsNotInitialized
         public MissionTask() { }
 
@@ -21,15 +21,19 @@ namespace Api.Database.Models
             Status = TaskStatus.NotStarted;
             IsarZoomDescription = copy.IsarZoomDescription;
             AnalysisTypes = copy.AnalysisTypes;
-            if (copy.Inspection is not null)
-            {
-                Inspection = new Inspection(copy.Inspection);
-            }
+            SensorType = copy.SensorType;
+            TargetPosition = copy.TargetPosition;
+            VideoDuration = copy.VideoDuration;
+            AcousticInspectionMetadata = copy.AcousticInspectionMetadata;
         }
 
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public string Id { get; set; }
+
+        [Required]
+        [MaxLength(200)]
+        public string IsarInspectionId { get; set; } = Guid.NewGuid().ToString();
 
         [Required]
         public int TaskOrder { get; set; }
@@ -44,6 +48,8 @@ namespace Api.Database.Models
         public Pose RobotPose { get; set; }
 
         public IList<AnalysisType>? AnalysisTypes { get; set; } = [];
+
+        private TaskStatus _status;
 
         [Required]
         public TaskStatus Status
@@ -77,16 +83,37 @@ namespace Api.Database.Models
 
         public IsarZoomDescription? IsarZoomDescription { get; set; }
 
-        public Inspection? Inspection { get; set; }
+        public SensorType SensorType { get; set; }
+
+        public Position TargetPosition { get; set; }
+
+        public float? VideoDuration { get; set; }
+
+        public AcousticInspectionMetadata? AcousticInspectionMetadata { get; set; }
 
         public string? ErrorDescription { get; set; }
 
-        public void UpdateWithIsarInfo(IsarTask isarTask)
+        public bool IsSupportedSensorType(IList<RobotCapabilitiesEnum> capabilities)
         {
-            if (isarTask.TaskType != IsarTaskType.ReturnToHome)
+            return SensorType switch
             {
-                Inspection?.UpdateWithIsarInfo(isarTask);
-            }
+                SensorType.Image => capabilities.Contains(RobotCapabilitiesEnum.take_image),
+                SensorType.ThermalImage => capabilities.Contains(
+                    RobotCapabilitiesEnum.take_thermal_image
+                ),
+                SensorType.Video => capabilities.Contains(RobotCapabilitiesEnum.take_video),
+                SensorType.ThermalVideo => capabilities.Contains(
+                    RobotCapabilitiesEnum.take_thermal_video
+                ),
+                SensorType.CO2Measurement => capabilities.Contains(
+                    RobotCapabilitiesEnum.take_co2_measurement
+                ),
+                SensorType.Audio => capabilities.Contains(RobotCapabilitiesEnum.record_audio),
+                SensorType.AcousticMeasurement => capabilities.Contains(
+                    RobotCapabilitiesEnum.take_acoustic_measurement
+                ),
+                _ => false,
+            };
         }
 
         public void UpdateStatus(IsarTaskStatus isarStatus)
@@ -103,24 +130,6 @@ namespace Api.Database.Models
                 _ => throw new ArgumentException($"ISAR Task status '{isarStatus}' not supported"),
             };
         }
-
-        public static string GetIsarInspectionTaskType()
-        {
-            return "inspection";
-        }
-
-        public TaskDefinition ToMissionTaskDefinition()
-        {
-            return new TaskDefinition
-            {
-                Index = this.TaskOrder,
-                TagId = this.TagId,
-                Description = this.Description,
-                RobotPose = this.RobotPose,
-                ZoomDescription = this.IsarZoomDescription,
-                AnalysisTypes = this.AnalysisTypes ?? [],
-            };
-        }
     }
 
     public enum TaskStatus
@@ -132,5 +141,107 @@ namespace Api.Database.Models
         Failed,
         Cancelled,
         Paused,
+    }
+
+    public enum SensorType
+    {
+        Image,
+        ThermalImage,
+        Video,
+        ThermalVideo,
+        Audio,
+        CO2Measurement,
+        AcousticMeasurement,
+    }
+
+    public enum AcousticDetectionType
+    {
+        [JsonStringEnumMemberName("leak")]
+        Leak,
+    }
+
+    [Owned]
+    public class AcousticInspectionMetadata : IValidatableObject
+    {
+        public const float MaxAcousticFrequencyHz = 100_000f;
+
+        [JsonConstructor]
+        public AcousticInspectionMetadata(
+            float frequencyFrom,
+            float frequencyTo,
+            float snrValueThreshold,
+            AcousticDetectionType detectionType
+        )
+        {
+            FrequencyFrom = frequencyFrom;
+            FrequencyTo = frequencyTo;
+            SnrValueThreshold = snrValueThreshold;
+            DetectionType = detectionType;
+        }
+
+        [Required]
+        [Range(0f, MaxAcousticFrequencyHz)]
+        public float FrequencyFrom { get; set; }
+
+        [Required]
+        [Range(0f, MaxAcousticFrequencyHz)]
+        public float FrequencyTo { get; set; }
+
+        [Required]
+        public float SnrValueThreshold { get; set; }
+
+        [Required]
+        public AcousticDetectionType DetectionType { get; set; }
+
+        public Roi? Roi { get; set; }
+
+        public AcousticInspectionMetadata(AcousticInspectionMetadata copy)
+            : this(copy.FrequencyFrom, copy.FrequencyTo, copy.SnrValueThreshold, copy.DetectionType)
+        {
+            Roi = copy.Roi is null ? null : new Roi(copy.Roi);
+        }
+
+        public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
+        {
+            if (FrequencyFrom >= FrequencyTo)
+            {
+                yield return new ValidationResult(
+                    $"{nameof(FrequencyFrom)} must be less than {nameof(FrequencyTo)}.",
+                    [nameof(FrequencyFrom), nameof(FrequencyTo)]
+                );
+            }
+        }
+    }
+
+    [Owned]
+    public class Roi
+    {
+        [JsonConstructor]
+        public Roi(int x, int y, int width, int height)
+        {
+            X = x;
+            Y = y;
+            Width = width;
+            Height = height;
+        }
+
+        [Required]
+        [Range(0, int.MaxValue)]
+        public int X { get; set; }
+
+        [Required]
+        [Range(0, int.MaxValue)]
+        public int Y { get; set; }
+
+        [Required]
+        [Range(1, int.MaxValue)]
+        public int Width { get; set; }
+
+        [Required]
+        [Range(1, int.MaxValue)]
+        public int Height { get; set; }
+
+        public Roi(Roi copy)
+            : this(copy.X, copy.Y, copy.Width, copy.Height) { }
     }
 }
